A few simple refactors
I started poking at the source again and thought I’d leave things a little better than I found them.

- Move init code out of `renderer/components/Meta.tsx` into `renderer/data/AppStorage.ts`
- Move random files from `renderer/` into `renderer/data` so it’s more obvious where to put more of that kind of thing
- Fix a bunch of warnings in `Meta.tsx`
