+++ src/renderer/components/config/BackupGroup.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import path from "path";
 
-import {getBackups, saveDir} from "../../utils";
+import {getBackups, saveDir} from "../../data/utils";
 import ControlGroup from "../sceneDetail/ControlGroup";
 import Modal from "../ui/Modal";
