+++ src/renderer/components/sceneDetail/EffectGroup.tsx
@@ -1,13 +1,13 @@
 import * as React from 'react';
-import {BT, HTF, VTF, ZF} from "../../const";
+import {BT, HTF, VTF, ZF} from "../../data/const";
 
 import SimpleOptionPicker from "../ui/SimpleOptionPicker";
 import ControlGroup from "./ControlGroup";
-import Scene from "../../Scene";
+import Scene from "../../data/Scene";
 import SimpleSliderInput from "../ui/SimpleSliderInput";
 import SimpleCheckbox from "../ui/SimpleCheckbox";
 import SimpleColorPicker from "../ui/SimpleColorPicker";
-import {SceneSettings} from "../../Config";
+import {SceneSettings} from "../../data/Config";
 import SimpleNumberInput from "../ui/SimpleNumberInput";
 
 export default class EffectGroup extends React.Component {
