+++ src/renderer/data/TIMING_FUNCTIONS.ts
