+++ src/renderer/data/en.ts
